enhnace: add MediaManger and lock iframe with css focusing.
