+++ apps/viewer.js
@@ -1,13 +1,8 @@
 import React, { useRef, useMemo } from "react";
 
-import {
-  Loaders,
-  ThreeCanvas,
-  PanoramaProjectionMesh,
-  Css3DObject,
-} from "../three";
+import { Loaders, ThreeCanvas, PanoramaProjectionMesh } from "../three";
 import useClick2AddWalls from "../hooks/useClick2AddWalls";
-import RawHTML from "../components/RawHTML";
+import MediaManager from "../components/MediaManager";
 
 const dev = process.env.NODE_ENV === "development";
 const Viewer = ({ data }) => {
@@ -40,11 +35,7 @@ const Viewer = ({ data }) => {
   return (
     <ThreeCanvas dev={dev} ref={threeRef}>
       <PanoramaProjectionMesh {...textureMeshProps} onLoad={onLoad} />
-      {data.media?.map(({ props, content }, index) => (
-        <Css3DObject {...props} key={index}>
-          <RawHTML content={content} />
-        </Css3DObject>
-      ))}
+      <MediaManager data={data.media} />
     </ThreeCanvas>
   );
 };
