--- components/RawHTML/index.js
@@ -1,13 +1,0 @@
-import { Wrapper } from "./styled";
-
-const RawHTML = ({ content }) => {
-  return (
-    <Wrapper
-      dangerouslySetInnerHTML={{
-        __html: content,
-      }}
-    />
-  );
-};
-
-export default RawHTML;
