--- components/RawHTML/styled.js
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-
-export const Wrapper = styled.div`
-  width: 100%;
-  height: 100%;
-`;
