+++ constant/media.js
@@ -1,0 +1,3 @@
+export const MEDIA = {
+  HTML: "HTML",
+};
